+++ scripts/macros/2024/classFeatures/bard/valor/combatInspiration.js
@@ -1,0 +1,5 @@
+export let combatInspiration = {
+    name: 'Combat Inspiration',
+    version: '1.2.11',
+    rules: 'modern'
+};